editoast: authz: define `authz::v2::infra` for organization
No code changes.
